Atualizado em 10/10/2017.

+++ app/Http/Controllers/Admin/UsersController.php
@@ -19,7 +19,8 @@ class UsersController extends Controller
      */
     public function index()
     {
-        $users = User::paginate();
+        Gate::authorize('r_users');
+        $users = User::paginate(10);
         return view('admin.users.index', compact('users'));
     }
 
@@ -46,7 +47,7 @@ class UsersController extends Controller
      */
     public function store(Request $request)
     {
-
+        Gate::authorize('crud_users');
         /** @var Form $form */
         $form = \FormBuilder::create(UserForm::class);
         if (!$form->isValid()){
@@ -72,6 +73,7 @@ class UsersController extends Controller
      */
     public function show(User $user)
     {
+        Gate::authorize('r_users');
         return view('admin.users.show', compact('user'));
     }
 
@@ -83,6 +85,7 @@ class UsersController extends Controller
      */
     public function edit(User $user)
     {
+        Gate::authorize('crud_users');
         $form = \FormBuilder::create(UserForm::class, [
             'url' => route('admin.users.update', ['user' => $user->id]),
             'method' => 'PUT',
@@ -100,6 +103,7 @@ class UsersController extends Controller
      */
     public function update(Request $request, User $user)
     {
+        Gate::authorize('crud_users');
         /** @var Form $form */
         $form = \FormBuilder::create(UserForm::class, [
             'data' => ['id' => $user->id]
@@ -124,6 +128,7 @@ class UsersController extends Controller
      */
     public function destroy(User $user)
     {
+        Gate::authorize('crud_users');
         try {
             $user->delete();
         } catch (Exception $e) {
